Added a delete listing button to all form pages.
In the middle of fixing the focus issue that happens on iOS safari. The issue is that when the form renders, the last InputStr component to render is set in focus when the body is what should be in focus instead. This is happening because of the forwardRef hack that is used in InputStr to get google places api to work with autocomplete

In the middle of breaking everything that has to do with the create listing page so that all the forms work with redux and forwardrefs are no longer needed